+++ tests/derive.rs
@@ -31,9 +31,9 @@ mod full_limbs {
     #[test]
     fn random_masking_does_not_overflow() {
         use ff::Field;
-        use rand::rngs::OsRng;
 
-        let _ = F384p::random(OsRng);
+        let _ = F384p::random(&mut rand::rng());
+        let _ = F384p::try_from_rng(&mut rand::rngs::OsRng).unwrap();
     }
 }
 
@@ -153,6 +153,5 @@ fn sqrt() {
 
     test(Fp::ZERO);
     test(Fp::ONE);
-    use rand::rngs::OsRng;
-    test(Fp::random(OsRng));
+    test(Fp::random(&mut rand::rng()));
 }